requirements.txt
django
djangorestframework
python-decouple
dj-database-url
